pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "air-drf-relation"
version = "0.7.0"
description = "Enhanced Django REST Framework serializers with advanced relation handling, queryset optimization, and type safety"
readme = "README.md"
authors = [{ name = "bubaley", email = "bubaley.fu@gmail.com" }]
maintainers = [{ name = "bubaley", email = "bubaley.fu@gmail.com" }]
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = [
    "django",
    "rest-framework",
    "drf",
    "serializer",
    "relations",
    "queryset-optimization",
    "preloading",
    "performance",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "django>=4.2.2",
    "djangorestframework>=3.14.0",
    "django-filter>=21.1",
    "dacite>=1.8.1",
    "djangorestframework-dataclasses>=1.2.0",
]

[project.urls]
Homepage = "https://github.com/bubaley/air-drf-relation"
Repository = "https://github.com/bubaley/air-drf-relation.git"
Documentation = "https://github.com/bubaley/air-drf-relation#readme"

[dependency-groups]
dev = [
    "coverage>=7.9.1",
    "pillow>=11.2.1",
    "pre-commit>=4.2.0",
    "ruff>=0.11.13",
    "django-environ>=0.12.0",
]

# Ruff configuration
[tool.ruff]
target-version = "py39"
line-length = 120
indent-width = 4
fix = true
format.quote-style = "single"
format.line-ending = "auto"
format.skip-magic-trailing-comma = false
lint.extend-select = ["I"]
lint.extend-per-file-ignores."core/settings/*.py" = ["F403", "F405"]
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Bandit configuration
[tool.bandit]
skips = ["B101"]
exclude_dirs = ["tests", "*/tests/*", "*/migrations/*"]
